\name{PhViD-package}
\alias{PhViD-package}
\alias{PhViD}
\docType{package}
\title{Pharmacovigilance signal detection methods extended to the multiple comparison setting.}

\description{
The \code{PhViD-package} proposes the main pharmacovigilance signal detection methods extended to the multiple comparison setting. Those functions can be used as standard R functions or through an user friendly interface (\code{PhViD.gui}). For the frequentist methods, the package requires the LBE procedure that can be downloaded from the Bioconductor website \url{http://bioconductor.org/packages/2.3/bioc/html/LBE.html}.
}
\details{
\tabular{ll}{
Package: \tab PhViD\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2009-02-11\cr
License: \tab GPL 2\cr
LazyLoad: \tab yes\cr
}
}
\author{
Ismail Ahmed & Antoine Poncet

Maintainer: Ismail Ahmed <ismail.ahmed@inserm.fr>
}
\references{
Ahmed I, Dalmasso C, Haramburu F, Thiessard F, Broet P, Tubert-Bitter P, False Discovery Rate Estimation for Frequentist Pharmacovigilance Signal Detection Methods, Biometrics, accepted. 

Ahmed I, Haramburu F, Fourrier-Reglat A, Thiessard F,  Kreft-Jais C, Miremont-Salame G, Begaud B, Tubert-Bitter P, Bayesian Pharmacovigilance Signal Detection Methods Revisited in a Multiple Comparison Setting, Statistics in Medicine, accepted. 

Bate A, Lindquist M, Edwards IR, Olsson S, Orre R, Lansner A, De Freitas RM. A Bayesian Neural Network Method for Adverse Drug Reaction Signal Generation European Journal of Clinical Pharmacology, 1998, 54, 315-321.  

Dalmasso C, Broet P, Moreau T (2005), A simple procedure for estimating the false discovery rate, Bioinformatics, Bioinformatics, 21: 660 - 668. 

DuMouchel W. Bayesian Data Mining in Large Frequency Tables, with an Application to the FDA Spontaneous Reporting System. The American Statistician. 1999, 53. 177-190. 

Evans SJ, Waller PC, Davis S, Use of Proportional Reporting Ratios (PRRs) for Signal Generation from Spontaneous Adverse Drug Reaction Reports Pharmacoepidemiology and Drug Safety, 2001, 10, 483-486. 

Noren, GN, Bate A, Orre R, Edwards IR, Extending the methods used to screen the WHO drug safety database towards analysis of complex associations and improved accuracy for rare events Statistics in Medicine, 2006, 25, 3740-3757. 

van Puijenbroek EP, Bate A, Leufkens HG, Lindquist M, Orre R, Egberts AC, A Comparison of Measures of Disproportionality for Signal Detection in Spontaneous Reporting Systems for Adverse Drug Reactions Pharmacoepidemiology and Drug Safety, 2002, 11, 3-1. 

}
\keyword{ htest }

